requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crontab==1.0.0
Django==4.1.7
django-extensions==3.2.1
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
numpy==1.24.2
pandas==1.5.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
